Fix Win64 build errors
TBR=abarth@chromium.org
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/61003006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@234931 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d1c8058f5704fa9f58c738e636858a0c8b32d425

+++ gin/converter.cc
@@ -89,7 +89,7 @@ Handle<Value> Converter<std::string>::ToV8(Isolate* isolate,
   return String::NewFromUtf8(isolate,
                              val.data(),
                              String::kNormalString,
-                             val.length());
+                             static_cast<uint32_t>(val.length()));
 }
 
 bool Converter<std::string>::FromV8(Handle<Value> val,
@@ -129,7 +129,7 @@ v8::Handle<v8::String> StringToSymbol(v8::Isolate* isolate,
   return String::NewFromUtf8(isolate,
                              val.data(),
                              String::kInternalizedString,
-                             val.length());
+                             static_cast<uint32_t>(val.length()));
 }
